Miscellaneous fixes from EndScreen integration

+++ core/src/test/java/bytemusketeers/heslingtonhustle/MovementTests.java
@@ -1,14 +1,14 @@
 package bytemusketeers.heslingtonhustle;
 
+import bytemusketeers.heslingtonhustle.entity.Player;
+import bytemusketeers.heslingtonhustle.map.GameMap;
 import bytemusketeers.heslingtonhustle.utils.CollisionHandler;
+import bytemusketeers.heslingtonhustle.utils.GameData;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.maps.tiled.TiledMap;
 import com.badlogic.gdx.maps.tiled.TmxMapLoader;
 import com.badlogic.gdx.math.Vector2;
-import bytemusketeers.heslingtonhustle.entity.Player;
-import bytemusketeers.heslingtonhustle.map.GameMap;
-import bytemusketeers.heslingtonhustle.utils.GameData;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -35,7 +35,7 @@ public class MovementTests {
     public void testMoveLeft() {
         // Simulating Player initialisation
         HeslingtonHustle game = new HeslingtonHustle();
-        float worldX = (float) game.screenWidth /2 - (float) game.screenHeight /2;
+        float worldX = (float) game.screenWidth / 2 - (float) game.screenHeight / 2;
         float worldY = 500;
         double normalizedSpeed = 200;
         TiledMap gameMap = new TmxMapLoader().load("map/MainMap.tmx");
@@ -45,7 +45,7 @@ public class MovementTests {
         CollisionHandler collisionHandler = new CollisionHandler(gameMap, tileSize, tileSize, spriteX, spriteY * 0.5f, 0.7f, 0.7f);
 
         // Calculating target coordinates when simulating movement
-        float targX = player.worldX - (float) (player.speed * Gdx.graphics.getDeltaTime());
+        float targX = player.worldX - (player.speed * Gdx.graphics.getDeltaTime());
         float targY = player.worldY;
 
         // Using a player's collision handler to derive the new position of a player after moving left
@@ -57,7 +57,7 @@ public class MovementTests {
     public void testMoveRight() {
         // Simulating Player initialisation
         HeslingtonHustle game = new HeslingtonHustle();
-        float worldX = (float) game.screenWidth /2 - (float) game.screenHeight /2;
+        float worldX = (float) game.screenWidth / 2 - (float) game.screenHeight / 2;
         float worldY = 500;
         double normalizedSpeed = 200;
         TiledMap gameMap = new TmxMapLoader().load("map/MainMap.tmx");
@@ -67,7 +67,7 @@ public class MovementTests {
         CollisionHandler collisionHandler = new CollisionHandler(gameMap, tileSize, tileSize, spriteX, spriteY * 0.5f, 0.7f, 0.7f);
 
         // Calculating target coordinates when simulating movement
-        float targX = player.worldX + (float) (player.speed * Gdx.graphics.getDeltaTime());
+        float targX = player.worldX + (player.speed * Gdx.graphics.getDeltaTime());
         float targY = player.worldY;
 
         // Using a player's collision handler to derive the new position of a player after moving left
@@ -79,7 +79,7 @@ public class MovementTests {
     public void testMoveUp() {
         // Simulating Player initialisation
         HeslingtonHustle game = new HeslingtonHustle();
-        float worldX = (float) game.screenWidth /2 - (float) game.screenHeight /2;
+        float worldX = (float) game.screenWidth / 2 - (float) game.screenHeight / 2;
         float worldY = 500;
         double normalizedSpeed = 200;
         TiledMap gameMap = new TmxMapLoader().load("map/MainMap.tmx");
@@ -90,7 +90,7 @@ public class MovementTests {
 
         // Calculating target coordinates when simulating movement
         float targX = player.worldX;
-        float targY = player.worldY + (float) (player.speed * Gdx.graphics.getDeltaTime());
+        float targY = player.worldY + (player.speed * Gdx.graphics.getDeltaTime());
 
         // Using a player's collision handler to derive the new position of a player after moving left
         Vector2 newPos = collisionHandler.adjustPos(player.worldX, player.worldY, targX, targY);
@@ -101,7 +101,7 @@ public class MovementTests {
     public void testMoveDown() {
         // Simulating Player initialisation
         HeslingtonHustle game = new HeslingtonHustle();
-        float worldX = (float) game.screenWidth /2 - (float) game.screenHeight /2;
+        float worldX = (float) game.screenWidth / 2 - (float) game.screenHeight / 2;
         float worldY = 500;
         double normalizedSpeed = 200;
         TiledMap gameMap = new TmxMapLoader().load("map/MainMap.tmx");
@@ -112,7 +112,7 @@ public class MovementTests {
 
         // Calculating target coordinates when simulating movement
         float targX = player.worldX;
-        float targY = player.worldY - (float) (player.speed * Gdx.graphics.getDeltaTime());
+        float targY = player.worldY - (player.speed * Gdx.graphics.getDeltaTime());
 
         // Using a player's collision handler to derive the new position of a player after moving left
         Vector2 newPos = collisionHandler.adjustPos(player.worldX, player.worldY, targX, targY);
